feat: display extension author in the extensions list
